requirements.txt
azure-functions==1.20.0
pyodbc==5.1.0
openai==1.46.0
